Move `monitorState` from `instances` to `monitor` reducer
